+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 ###
 - Fix cross-instance action queue bug
+- Fix unresolved promise bug within action flushing algorithm
 
 ## 0.1.1 (2021-6-15)
 ### Fixed